swapped index pages
